aa-integration-backend/ui-connector/requirements.txt
aiohttp==3.7.4.post0
async-timeout==3.0.1
attrs==21.2.0
bidict==0.21.3
cachetools==4.2.4
certifi==2021.10.8
chardet==4.0.0
charset-normalizer==2.0.7
click==8.0.1
Flask==2.0.1
Flask-Cors==3.0.10
Flask-SocketIO==5.1.1
google-api-core==2.8.0
google-auth==2.3.0
google-cloud-dialogflow==2.9.1
google-cloud-secret-manager==2.10.0
googleapis-common-protos==1.53.0
grpcio==1.41.0
gunicorn==20.1.0
h11==0.12.0
idna==3.2
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
multidict==5.1.0
packaging==21.0
proto-plus==1.19.5
protobuf==3.18.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyJWT==2.3.0
pyparsing==2.4.7
python-engineio==4.2.1
python-socketio==5.4.0
redis==3.5.3
requests==2.26.0
rsa==4.7.2
simple-websocket==0.4.0
six==1.16.0
typing-extensions==3.10.0.2
urllib3==1.26.7
Werkzeug==2.0.1
wsproto==1.0.0
yarl==1.6.3